replace the optimizer/smoothing steps by a scheme that can provide an osg-specific mesh file
For each mesh file, the code now checks for a file called <filename>.osg.
This file gets loaded instead of the original mesh if present. This allows
to run the osg optimizers offline (thus reducing the amount of time
spent loading quite a bit), and it also allows to tune the post-processing
of meshes based on each individual meshes needs, instead of relying on
having a default that works for everything.